Refactoring

+++ Delivery/internal/repositories/models/OrderProduct.go
@@ -1,9 +1,9 @@
 package models
 
 type OrderProduct struct {
-	Id        int     `json:"id,omitempty"`
-	OrderId   int     `json:"orderId,order_id,omitempty"`
-	ProductId int     `json:"productId,product_id,omitempty"`
-	Count     int     `json:"count,counter,omitempty"`
+	Id        int64   `json:"id,omitempty"`
+	OrderId   int64   `json:"orderId,order_id,omitempty"`
+	ProductId int64   `json:"productId,product_id,omitempty"`
+	Count     int64   `json:"count,counter,omitempty"`
 	Price     float64 `json:"price,omitempty"`
 }
