+++ Delivery/internal/repositories/models/Product.go
@@ -1,9 +1,9 @@
 package models
 
 type Product struct {
-	Id          int      `json:"id"`
+	Id          int64    `json:"id"`
 	Name        string   `json:"name"`
-	MenuId      int      `json:"menuId"`
+	MenuId      int64    `json:"menuId"`
 	Price       float64  `json:"price"`
 	Image       string   `json:"image"`
 	Type        string   `json:"type"`
